decouple SDL dependencies on CPU and GPU by moving more to /alpaka
