+++ RecoTracker/LST/plugins/alpaka/LSTModulesDevESProducer.cc
@@ -5,15 +5,24 @@
 #include "HeterogeneousCore/AlpakaInterface/interface/config.h"
 #include "HeterogeneousCore/AlpakaInterface/interface/memory.h"
 
-#include "RecoTracker/LST/interface/LSTModulesDev.h"
 #include "RecoTracker/Record/interface/TrackerRecoGeometryRecord.h"
 
 // LST includes
+#include <SDL/Module.h>
 #include <SDL/ModuleMethods.h>
 #include <SDL/LST.h>
 
+namespace cms::alpakatools {
+  template <>
+  struct CopyToDevice<SDL::modulesBuffer<alpaka_common::DevHost>> {
+    template <typename TQueue>
+    static auto copyAsync(TQueue& queue, SDL::modulesBuffer<alpaka_common::DevHost> const& srcData) {
+      return SDL::modulesBuffer<alpaka::Dev<TQueue>>(queue, srcData);
+    }
+  };
+} //namespace cms::alpakatools
 namespace ALPAKA_ACCELERATOR_NAMESPACE {
-  
+
 class LSTModulesDevESProducer : public ESProducer {
 public:
   LSTModulesDevESProducer(const edm::ParameterSet &iConfig);
