--- RecoTracker/LST/src/ES_ModulesDev.cc
@@ -1,3 +1,0 @@
-#include "RecoTracker/LST/interface/LSTModulesDev.h"
-#include "FWCore/Utilities/interface/typelookup.h"
-TYPELOOKUP_DATA_REG(SDL::modulesBuffer<alpaka_common::DevHost>);
