+++ RecoTracker/LST/src/alpaka/ES_ModulesDev.cc
@@ -1,3 +1,7 @@
-#include "RecoTracker/LST/interface/LSTModulesDev.h"
+#include <SDL/Module.h>
 #include "HeterogeneousCore/AlpakaCore/interface/alpaka/typelookup.h"
+
+// Temporary hack: The DevHost instantiation is needed in both CPU and GPU plugins,
+// whereas the (non-host-)Device instantiation only in the GPU plugin
+TYPELOOKUP_DATA_REG(SDL::modulesBuffer<alpaka_common::DevHost>);
 TYPELOOKUP_ALPAKA_TEMPLATED_DATA_REG(SDL::modulesBuffer);
